move properties and funcs that are easily wrapped
